[TASK] Fix phpstan checkFunctionArgumentTypes errors in ext:fluid
This patch fixes incompatible type usage in function arguments
and is preparatory work for introducing native type hints and
strict mode in all core files.

Resolves: #92167
Releases: master, 10.4
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/65911
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Daniel Goerz <daniel.goerz@posteo.de>
Tested-by: Benni Mack <benni@typo3.org>

+++ typo3/sysext/fluid/Classes/ViewHelpers/Format/DateViewHelper.php
@@ -155,7 +155,7 @@ class DateViewHelper extends AbstractViewHelper
 
         if (!$date instanceof \DateTimeInterface) {
             try {
-                $base = $base instanceof \DateTimeInterface ? $base->format('U') : strtotime((MathUtility::canBeInterpretedAsInteger($base) ? '@' : '') . $base);
+                $base = $base instanceof \DateTimeInterface ? (int)$base->format('U') : (int)strtotime((MathUtility::canBeInterpretedAsInteger($base) ? '@' : '') . $base);
                 $dateTimestamp = strtotime((MathUtility::canBeInterpretedAsInteger($date) ? '@' : '') . $date, $base);
                 $date = new \DateTime('@' . $dateTimestamp);
                 $date->setTimezone(new \DateTimeZone(date_default_timezone_get()));
@@ -165,7 +165,7 @@ class DateViewHelper extends AbstractViewHelper
         }
 
         if (strpos($format, '%') !== false) {
-            return strftime($format, $date->format('U'));
+            return strftime($format, (int)$date->format('U'));
         }
         return $date->format($format);
     }
